Move scopedFilter back to ip.go
