+++ pkg/repository/repository.go
@@ -7,7 +7,6 @@ import (
 	asyncclient "github.com/metal-stack/metal-apiserver/pkg/async/client"
 	"github.com/metal-stack/metal-apiserver/pkg/db/generic"
 	"github.com/metal-stack/metal-apiserver/pkg/db/metal"
-	"github.com/metal-stack/metal-apiserver/pkg/db/queries"
 
 	adminv2 "github.com/metal-stack/api/go/metalstack/admin/v2"
 	apiv2 "github.com/metal-stack/api/go/metalstack/api/v2"
@@ -148,16 +147,4 @@ func (r *Store) FilesystemLayout() FilesystemLayout {
 	return &filesystemLayoutRepository{
 		r: r,
 	}
-}
-
-func (r *ipRepository) scopedFilters(filter generic.EntityQuery) []generic.EntityQuery {
-	var qs []generic.EntityQuery
-	r.r.log.Info("scopedFilters", "scope", r.scope)
-	if r.scope != nil {
-		qs = append(qs, queries.IpProjectScoped(r.scope.projectID))
-	}
-	if filter != nil {
-		qs = append(qs, filter)
-	}
-	return qs
 }
